Wallet/src/com/cg/customer/bean/Customer.java
package com.cg.customer.bean;

public class Customer {
	String name;
	String mobno;
	Wallet wallet;
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public String getMobno() {
		return mobno;
	}
	public void setMobno(String mobno) {
		this.mobno = mobno;
	}
	public Wallet getWallet() {
		return wallet;
	}
	public void setWallet(Wallet wallet) {
		this.wallet = wallet;
	}
	@Override
	public String toString() {
		return "Customer [name=" + name + ", mobno=" + mobno + ", wallet=" + wallet + "]";
	}
	
}